save snapshot with cast, not call
Interestingly the tests still pass without any changes. It seems that
this is because of the transaction effectively serlialising the test. I
think there's some confirmation of this in the Elixir Fountain podcast,
episode 23 with Jose Valim.
https://soundcloud.com/elixirfountain/elixir-fountain-2016-02-05-jose-valim

From what I can tell from the podcast, this will still work with future
db tests that will be parallisable.